# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PhotoOrganizer API",
    "description": "PhotoOrganizer image organizer",
    "license": {
      "name": "MIT",
      "url": "https://github.com/KKDad/PhotoOrganizerAPI/LICENSE"
    },
    "version": "v0.0.1"
  },
  "externalDocs": {
    "description": "PhotoOrganizer Documentation",
    "url": "https://github.com/KKDad/PhotoOrganizerAPI/docs"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": [
        "bearer-token"
      ],
      "apiKey": [
        "api-key"
      ]
    }
  ],
  "paths": {
    "/api/v1/accounts/{id}": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "summary": "Get a account by its id",
        "operationId": "getaccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid id supplied",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          },
          "200": {
            "description": "Found the account",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "account-controller"
        ],
        "summary": "Update a account",
        "operationId": "updateaccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountInfoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid account supplied",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          },
          "200": {
            "description": "Updated the account",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          },
          "404": {
            "description": "account not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "account-controller"
        ],
        "summary": "Delete a account",
        "operationId": "deleteaccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid id supplied"
          },
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Deleted the account"
          },
          "404": {
            "description": "account not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/auth/authenticate": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "authenticateAndGetToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "summary": "Get all accounts",
        "operationId": "getAllaccounts",
        "responses": {
          "404": {
            "description": "Accounts not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountInfoDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountInfoDto"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Found the accounts",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountInfoDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountInfoDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "account-controller"
        ],
        "summary": "Create a new account",
        "operationId": "createaccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountInfoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid account supplied",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          },
          "201": {
            "description": "Created the account",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/welcome": {
      "get": {
        "tags": [
          "home-controller"
        ],
        "operationId": "welcome",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/userProfile": {
      "get": {
        "tags": [
          "home-controller"
        ],
        "operationId": "userProfile",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/adminProfile": {
      "get": {
        "tags": [
          "home-controller"
        ],
        "operationId": "adminProfile",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/": {
      "get": {
        "tags": [
          "home-controller"
        ],
        "operationId": "getHome",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccountInfoDto": {
        "required": [
          "email",
          "roles",
          "username"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john@stapledon.local"
          },
          "password": {
            "type": "string",
            "description": "Password. Only used when creating or updating a user. Never returned from the server. Leave blank when updating a user to keep the existing password.",
            "example": "password"
          },
          "username": {
            "type": "string",
            "description": "Username used to log in",
            "example": "kkdad"
          },
          "firstName": {
            "type": "string",
            "description": "users firstname",
            "nullable": true,
            "example": "Johnathan"
          },
          "lastName": {
            "type": "string",
            "description": "users surname",
            "nullable": true,
            "example": "McDoe"
          },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "description": "Assigned Roles",
            "nullable": true,
            "example": "ROLE_ADMIN",
            "items": {
              "type": "string",
              "description": "Assigned Roles",
              "nullable": true,
              "example": "ROLE_ADMIN"
            }
          }
        },
        "description": "Account information"
      },
      "AuthRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "description": "Authentication request"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "description": "API Key Authentication",
        "name": "api-key",
        "in": "header"
      }
    }
  }
}